CoreSightTarget: moved under coresight; added SoCTarget superclass.
- SoCTarget represents an SoC.
- CoreSightTarget adds the Arm CoreSight specifics of the SoC.
- Updated all references to CoreSightTarget.
